fix chunking
