+++ services/apps/premium/members_enrichment_worker/src/utils/common.ts
@@ -100,3 +100,11 @@ export function normalizeAttributes(
 
   return normalized
 }
+
+export function chunkArray<T>(array: T[], chunkSize: number): T[][] {
+  const chunks = []
+  for (let i = 0; i < array.length; i += chunkSize) {
+    chunks.push(array.slice(i, i + chunkSize))
+  }
+  return chunks
+}
